question_generation_service/pyproject.toml
[project]
name = "question-generation-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "langchain==0.3.27",
    "langchain-core==0.3.74",
    "langchain-openai==0.3.29",
    "pika>=1.3.2",
    "python-dotenv>=1.1.1",
]
